Move all modes into separate module
