+++ src/buffer.rs
@@ -5,8 +5,8 @@ use std::path::{Path, PathBuf};
 
 use super::byte_rope::*;
 use super::history::History;
-use super::mode::*;
-use super::selection::*;
+use crate::modes::mode::DirtyBytes;
+use crate::selection::{SelRegion, Selection};
 
 #[derive(Debug, PartialEq, Clone, Copy, Eq, Hash)]
 pub enum OverflowSelectionStyle {
