+++ src/modes/command.rs
@@ -1,14 +1,17 @@
-use super::buffer::*;
-use super::keymap::*;
-use super::mode::*;
-use super::modes::normal::Normal;
-use super::modes::quitting;
-use crossterm::event::{Event, KeyCode, KeyEvent, KeyModifiers};
-use lazy_static::lazy_static;
-use maplit::hashmap;
 use std::borrow::Cow;
 use std::collections::HashMap;
 use std::fs;
+
+use crossterm::event::{Event, KeyCode, KeyEvent, KeyModifiers};
+use lazy_static::lazy_static;
+use maplit::hashmap;
+
+use crate::keymap::KeyMap;
+use crate::modes::{
+    mode::{Mode, ModeTransition},
+    normal::Normal,
+};
+use crate::Buffers;
 
 pub struct Command {
     pub command: String,
@@ -40,6 +43,8 @@ fn default_maps() -> KeyMap<Action> {
 
 mod cmd {
     use super::*;
+    use crate::modes::mode::DirtyBytes;
+    use crate::modes::quitting;
 
     pub fn quit(buf: &mut Buffers, _: &str) -> ModeTransition {
         if buf.iter().any(|x| x.dirty && x.path.is_some()) {
