+++ src/modes/jumpto.rs
@@ -1,20 +1,21 @@
 use std::borrow::Cow;
 use std::collections::HashMap;
 
-use super::buffer::*;
-use super::keymap::*;
-use super::mode::*;
-use super::modes::normal::Normal;
-use super::selection::*;
-
 use crossterm::event::{Event, KeyCode, KeyEvent, KeyModifiers};
+use lazy_static::lazy_static;
+
+use crate::keymap::KeyMap;
+use crate::modes::{
+    mode::{Mode, ModeTransition},
+    normal::Normal,
+};
+use crate::selection::Direction;
+use crate::Buffers;
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub struct JumpTo {
     pub extend: bool,
 }
-
-use lazy_static::lazy_static;
 
 fn default_maps() -> KeyMap<Direction> {
     KeyMap {
